requirements.txt
# Trading Platform Requirements

# Core dependencies
python-dotenv==1.2.1
psycopg2-binary==2.9.11
pandas==2.2.3
numpy>=2.2.6  # Updated for pandas-ta compatibility
yfinance==0.2.50

# Technical analysis
TA-Lib==0.4.28
pandas-ta>=0.3.14b

# Backtesting (optional, install when needed)
# backtrader==1.9.78.123
# backtesting==0.3.3

# Visualization
matplotlib==3.10.0
plotly==5.24.1
mplfinance==0.12.10b0

# Notifications (optional)
python-telegram-bot==21.10

# Development tools
pytest==8.3.4
pytest-cov==6.0.0
black==24.10.0
flake8==7.1.1

# Jupyter (for analysis)
jupyter==1.1.1
ipykernel==6.29.5
google-generativeai>=0.3.0
feedparser>=6.0.0
openai>=1.0.0
